refactor: modify lsj import to use html
